+++ packages/stream_video/lib/src/sfu/data/models/sfu_pin.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 
-class SfuPin with EquatableMixin {
-  SfuPin({
+class SfuPin extends Equatable {
+  const SfuPin({
     required this.userId,
     required this.sessionId,
   });